requirements.txt
streamlit
qrcode[pil]
fpdf
pillow
pandas